Code refactor for hibernate session handling

+++ src/main/java/com/cts/fsd/dao/TaskDetailsDao.java
@@ -10,7 +10,7 @@ public interface TaskDetailsDao {
 
 	void updateTask(TaskDetails task);
 
-	TaskDetails searchTask(long taskId);
+	TaskDetails searchTask(Integer taskId);
 
 	void deleteTask(TaskDetails task);
